+++ upgrades/0.0.94-increase-argocd-memlimit/.golangci.yml
