+++ upgrades/0.0.94-increase-argocd-memlimit/pkg/lib/commonerrors/api.go
