+++ upgrades/0.0.94-increase-argocd-memlimit/pkg/lib/kubectl/types.go
@@ -3,7 +3,7 @@ package kubectl
 import (
 	"errors"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.93-increase-argocd-memlimit/pkg/lib/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.94-increase-argocd-memlimit/pkg/lib/logger"
 	"k8s.io/client-go/kubernetes"
 )
