+++ upgrades/0.0.94-increase-argocd-memlimit/upgrade.go
@@ -3,8 +3,8 @@ package main
 import (
 	"fmt"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.93-increase-argocd-memlimit/pkg/lib/cmdflags"
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.93-increase-argocd-memlimit/pkg/memlimit"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.94-increase-argocd-memlimit/pkg/lib/cmdflags"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.94-increase-argocd-memlimit/pkg/memlimit"
 )
 
 func upgrade(context Context, flags cmdflags.Flags) error {
